Move map to ES modules
