+++ map/api/models/alerts.model.mongodb.js
@@ -1,4 +1,4 @@
-module.exports = function (app, options) {
+export default function (app, options) {
   const db = options.db || app.db
   options.Model = db.collection('alerts')
   options.Model.createIndex({ geometry: '2dsphere' })
