+++ map/api/models/features.model.mongodb.js
@@ -1,4 +1,4 @@
-module.exports = function (app, options) {
+export default function (app, options) {
   const db = options.db || app.db
   options.Model = db.collection(options.collection)
   options.Model.createIndex({ geometry: '2dsphere' })
