+++ package.json
@@ -31,11 +31,9 @@
     "release:minor": "npm version minor && npm publish --access=public",
     "release:major": "npm version major && npm publish --access=public",
     "changelog": "github_changelog_generator -u kalisio -p kdk && git add CHANGELOG.md && git commit -am \"Updating changelog\"",
-    "compile:map": "shx rm -rf lib/map && babel -d lib/map map/ -s --copy-files",
     "compile:tests:client": "shx rm -rf lib/test/client && babel -d lib/test/client test/client -s --copy-files",
     "compile": "npm-run-all --serial compile:**",
     "watch:core:vue": "cpx core/**/*.vue lib/core --watch",
-    "watch:map:js": "babel --watch -d lib/map map/ --copy-files",
     "watch:map:vue": "cpx map/**/*.vue lib/map --watch",
     "watch:test:client": "cpx \"test/client/**/*.js\" lib/test/client --watch",
     "watch": "npm-run-all --parallel watch:**",
@@ -45,7 +43,7 @@
     "lint:test:client": "standard test/client/core/**/*.js test/client/map/**/*.js --fix",
     "lint": "npm-run-all --serial lint:**",
     "mocha:core": "cross-env NODE_CONFIG_DIR=./test/api/core/config/ mocha test/api/core/**/*.test.js --exit",
-    "mocha:map": "cross-env NODE_CONFIG_DIR=./test/api/map/config/ mocha test/api/map/**/*.test.js --exit --require babel-core/register",
+    "mocha:map": "cross-env NODE_CONFIG_DIR=./test/api/map/config/ mocha test/api/map/**/*.test.js --exit",
     "mocha": "npm-run-all --serial mocha:**",
     "coverage": "nyc npm run mocha",
     "test": "npm run compile && npm run lint && npm run coverage"
@@ -142,6 +140,9 @@
     "@feathersjs/authentication-client": "^1.0.2",
     "@kalisio/feathers-distributed": "^1.0.4",
     "@turf/intersect": "^6.1.3",
+    "@weacast/core": "^2.0.0",
+    "@weacast/gfs": "^2.0.0",
+    "@weacast/probe": "^2.0.0",
     "babel-cli": "^6.24.1",
     "babel-eslint": "^10.0.3",
     "babel-plugin-add-module-exports": "^0.2.1",
@@ -168,9 +169,6 @@
     "pngjs": "^5.0.0",
     "puppeteer": "^10.2.0",
     "shx": "^0.3.2",
-    "standard": "^14.0.0",
-    "weacast-core": "^1.4.0",
-    "weacast-gfs": "^1.5.0",
-    "weacast-probe": "^1.4.0"
+    "standard": "^14.0.0"
   }
 }
