+++ test/api/map/hooks.test.js
@@ -1,6 +1,8 @@
-import chai, { util, expect } from 'chai'
+import chai from 'chai'
 import chailint from 'chai-lint'
-import { hooks } from '../../../map/api'
+import { hooks } from '../../../map/api/index.js'
+
+const { util, expect } = chai
 
 describe('map:hooks', () => {
   before(() => {
